% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C2F}
\alias{C2F}
\title{Celsius conversion}
\usage{
C2F(TC)
}
\arguments{
\item{TC}{is the temp in Celsius}
}
\value{
temp in Fahrenheit
}
\description{
Celsius conversion
}
\examples{
temp1 <- C2F(22)
temp2 <- C2F(c(-2, 12, 23))
}
