% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F2C}
\alias{F2C}
\title{Fahrenheit conversion}
\usage{
F2C(TF)
}
\arguments{
\item{TF}{is the temp in Fahrenheit}
}
\value{
temp in Celsius
}
\description{
Fahrenheit conversion
}
\examples{
temp1 <- F2C(50)
temp2 <- F2C(c(50,63,23))
}
